partial: new testnet wip 1 (#16)
* wip: start of new testnet

* tweak: add dom to tsconfig.json

* Update package.json

* chore: remove btc wrapping initial test

* tweak: modifications to hive stream lib

* fix: null val, remove old

* feat (update.sh): update if git is diff

* partial: getWitnessesAtBlock

* fix: add back bitcoin-utils, code cleanup

* fix: btc scripts

* misc

* feat: BLS DID lib

* partial: btc unwrap scripts

* feat (smart contract VM): support base58 encode/decode

* wip: unwrap code + did decode + other

+++ update.sh
@@ -3,11 +3,23 @@ set -e
 
 updateCode=$(git pull);
 
-if [[ "$updateCode" == "Already up to date." ]]
-then
-  echo $updateCode
-  exit
-fi
+gitCommit=$(git rev-parse HEAD)
+
+
+# if [[ "$updateCode" == "Already up to date." ]]
+# then
+#   echo $updateCode
+#   exit
+# fi
+{ # try
+
+    deployedGitCommit=`cat data/git-flag`
+    #save your output
+
+} || { # catch
+    deployedGitCommit="0"
+}
+echo "$gitCommit $deployedGitCommit"
 
 indexVersion=`cat deploy/index-flag`
 { # try
@@ -20,19 +32,30 @@ indexVersion=`cat deploy/index-flag`
 }
 echo "$indexVersion $indexVersionDeployed"
 
-docker-compose down
-
-sleep 15; # Ensure safe shutdown
-
-docker-compose build
-
-
-if [ "$indexVersion" -ne "$indexVersionDeployed" ];
+if [ "$gitCommit" != "$deployedGitCommit" ];
 then
-  echo "Need to reset database"
-  rm -r data/*
-  echo $indexVersion > data/index-flag
+  docker-compose down
+
+  sleep 15; # Ensure safe shutdown
+
+  docker-compose build
+
+
+  if [ "$indexVersion" -ne "$indexVersionDeployed" ];
+  then
+    echo "Need to reset database"
+    { # try
+      rm -r data/*
+    } || { # catch
+        indexVersionDeployed="0"
+    }
+    
+    echo $indexVersion > data/index-flag
+  fi
+
+
+  docker-compose up -d
+
+  echo $deployedGitCommit > data/git-flag
 fi
 
-
-docker-compose up -d
